Add Inspect function to avoid querying Datadog for deletes and general cleanup.
